--- src/main/java/com/prgrms/amabnb/auth/jwt/JwtTokenProvider.java
@@ -1,66 +1,0 @@
-package com.prgrms.amabnb.auth.jwt;
-
-import java.nio.charset.StandardCharsets;
-import java.util.Date;
-import java.util.UUID;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
-
-@Component
-public class JwtTokenProvider {
-
-    @Value("${jwt.issuer}")
-    private String issuer;
-    @Value("${jwt.secret-key}")
-    private String secretKey;
-    @Value("${jwt.access-token.expire-length}")
-    private long accessTokenValidityInMilliseconds;
-    @Value("${jwt.refresh-token.expire-length}")
-    private long refreshTokenValidityInMilliseconds;
-
-    public String createAccessToken(String payload) {
-        return createToken(payload, accessTokenValidityInMilliseconds);
-    }
-
-    public String createRefreshToken() {
-        String payload = UUID.randomUUID().toString();
-        return createToken(payload, refreshTokenValidityInMilliseconds);
-    }
-
-    private String createToken(String payload, long expireLength) {
-        Claims claims = Jwts.claims()
-            .setSubject(payload);
-        Date now = new Date();
-        Date expiredDate = new Date(now.getTime() + expireLength);
-
-        return Jwts.builder()
-            .setIssuer(issuer)
-            .setClaims(claims)
-            .setIssuedAt(now)
-            .setExpiration(expiredDate)
-            .signWith(Keys.hmacShaKeyFor(secretKey.getBytes(StandardCharsets.UTF_8)))
-            .compact();
-    }
-
-    public String getPayload(String token) {
-        return Jwts.parserBuilder()
-            .setSigningKey(Keys.hmacShaKeyFor(secretKey.getBytes(StandardCharsets.UTF_8)))
-            .build()
-            .parseClaimsJws(token)
-            .getBody()
-            .getSubject();
-    }
-
-    public void validateToken(String token) {
-        Jwts.parserBuilder()
-            .setSigningKey(Keys.hmacShaKeyFor(secretKey.getBytes(StandardCharsets.UTF_8)))
-            .build()
-            .parseClaimsJws(token);
-    }
-
-}
